+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Development
 
 - Interpolation/Summary: Now the queried point can be an existing station laying on the border of the polygon that it's
   being checked against
+- Geo: Change function signatures to use latlon tuple instead of latitude and longitude
 
 0.49.0 (28.11.2022)
 *******************